requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
pymongo==4.6.0
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
yt-dlp==2023.11.16
cryptography==41.0.8
pyOpenSSL==23.3.0 
